redirector/requirements.txt
bottle==0.12.18 \
    --hash=sha256:0819b74b145a7def225c0e83b16a4d5711fde751cd92bae467a69efce720f69e \
    --hash=sha256:43157254e88f32c6be16f8d9eb1f1d1472396a4e174ebd2bf62544854ecf37e7
python-decouple==3.3 \
    --hash=sha256:55c546b85b0c47a15a47a4312d451a437f7344a9be3e001660bccd93b637de95
gunicorn==20.0.4 \
    --hash=sha256:1904bb2b8a43658807108d59c3f3d56c2b6121a701161de0ddf9ad140073c626 \
    --hash=sha256:cd4a810dd51bf497552cf3f863b575dabd73d6ad6a91075b65936b151cbf4f9c
